Add initial Dockerfile
This is a simple take on a docker build that uses the rakudo base container.

Test from the root of the directory like this

    docker build -t localhost/rakuast-rakudoc-render:latest .

The `podman` executable can be used instead of `docker`, as well.

Details on usage are TBD, but it will involve users mounting their source
code inside as a volume mount.